// src/libs/api/schemas/processingResult.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { Point } from './point';
import type { PedestrianRouteResponse } from './pedestrianRouteResponse';

export interface ProcessingResult {
  /** Distance to reach destination (meters) */
  distance: number;
  /** 각 라벨의 개수 */
  labelCount: number[];
  /** Processed points */
  points: Point[];
  /** Pedestrian route response */
  road: PedestrianRouteResponse;
  /** Time to reach destination (seconds) */
  time: number;
  /** Walkablity index */
  walkablityIndex: number;
}